Animated: Hoist Non-Lifecycle Logic from `useAnimatedPropsLifecycle` (#48877)
Summary:

This is a minor refactor to hoist logic that is not actually specific to the `AnimatedProps` lifecycle out of the hook named `useAnimatedPropsLifecycle`.

This will make it easier to iterate on the implementation of `useAnimatedPropsLifecycle` using a feature flag in a subsequent diff.

This has no behavior change.

Changelog:
[Internal]

Differential Revision: D68516034